+++ pkg/resourceprovider/resourceprovider.go
@@ -19,6 +19,7 @@ import (
 	"github.com/lilypad-tech/lilypad/pkg/web3"
 	"github.com/lilypad-tech/lilypad/pkg/web3/bindings/pow"
 	"github.com/rs/zerolog/log"
+	"go.opentelemetry.io/otel/trace"
 )
 
 // this configures the resource offers we will keep track of
@@ -82,9 +83,9 @@ func NewResourceProvider(
 	options ResourceProviderOptions,
 	web3SDK *web3.Web3SDK,
 	executor executor.Executor,
-	telemetry system.Telemetry,
+	tracer trace.Tracer,
 ) (*ResourceProvider, error) {
-	controller, err := NewResourceProviderController(options, web3SDK, executor, telemetry)
+	controller, err := NewResourceProviderController(options, web3SDK, executor, tracer)
 	if err != nil {
 		return nil, err
 	}
@@ -118,7 +119,9 @@ func (resourceProvider *ResourceProvider) StartMineLoop(ctx context.Context) cha
 
 	taskCh := make(chan Task)
 	resourceProvider.controller.web3Events.Pow.SubscribenewPowRound(func(newPowRound pow.PowNewPowRound) {
+
 		_, challenge, err := resourceProvider.web3SDK.GetGenerateChallenge(ctx, nodeId)
+
 		if err != nil {
 			log.Err(err).Msgf("Unable to fetch challenge")
 			return
@@ -127,11 +130,19 @@ func (resourceProvider *ResourceProvider) StartMineLoop(ctx context.Context) cha
 		log.Info().Msgf("Receive a new pow signal challenge hex %s, difficulty %s", "0x"+hex.EncodeToString(challenge.Challenge[:]), challenge.Difficulty.String())
 
 		difficulty, _ := uint256.FromBig(challenge.Difficulty)
+		uuid := uuid.New()
+
+		err = PostCard(uuid.String(), "0x"+hex.EncodeToString(challenge.Challenge[:]), challenge.Difficulty.String())
+		if err != nil {
+			log.Err(err).Msgf("Unable to post card")
+		}
+
 		taskCh <- Task{
-			Id:         uuid.New(),
+			Id:         uuid,
 			Challenge:  challenge.Challenge,
 			Difficulty: difficulty,
 		}
+
 	})
 
 	submitWork := func(nonce *big.Int, hashrate float64) {